+++ nwt-cinema-movies/src/main/java/ba/unsa/etf/nwtcinemamovies/controllers/RolesController.java
@@ -1,7 +1,7 @@
 package ba.unsa.etf.nwtcinemamovies.controllers;
 
-import ba.unsa.etf.nwtcinemamovies.models.UserRole;
-import ba.unsa.etf.nwtcinemamovies.services.UserRoleService;
+import ba.unsa.etf.nwtcinemamovies.models.Role;
+import ba.unsa.etf.nwtcinemamovies.services.RoleService;
 import ba.unsa.etf.nwtcinemamovies.utils.JSONConverter;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestBody;
@@ -12,36 +12,36 @@ import org.springframework.web.bind.annotation.RestController;
 import javax.transaction.Transactional;
 
 @RestController
-@RequestMapping(value = "userRoles", produces = "application/json")
-public class UserRoles extends AbstractController<UserRoleService> {
+@RequestMapping(value = "nwt_cinema/movies/roles", produces = "application/json")
+public class RolesController extends AbstractController<RoleService> {
 
 	@Transactional
 	@RequestMapping(value = "create", method = RequestMethod.POST)
-	public void create(@RequestBody final UserRole role) {
+	public void create(@RequestBody final Role role) {
 		service.save(role);
 	}
 
 	@Transactional
 	@RequestMapping(value = "update", method = RequestMethod.POST)
-	public String update(@RequestBody final UserRole role) {
+	public String update(@RequestBody final Role role) {
 		return JSONConverter.toJSON(service.update(role));
 	}
 
 	@Transactional
 	@RequestMapping(value = "{roleId}", method = RequestMethod.GET)
 	public String findById(@PathVariable("roleId") final Integer roleId) {
-		return JSONConverter.toJSON(service.findById(UserRole.class, roleId));
+		return JSONConverter.toJSON(service.findById(Role.class, roleId));
 	}
 
 	@Transactional
 	@RequestMapping(value = "findAll", method = RequestMethod.GET)
 	public String findAll() {
-		return JSONConverter.toJSON(service.findAll(UserRole.class));
+		return JSONConverter.toJSON(service.findAll(Role.class));
 	}
 
 	@Transactional
 	@RequestMapping(value = "delete", method = RequestMethod.DELETE)
-	public void delete(@RequestBody final UserRole role) {
+	public void delete(@RequestBody final Role role) {
 		service.delete(role);
 	}
 }
