+++ nwt-cinema-movies/src/main/java/ba/unsa/etf/nwtcinemamovies/controllers/UserRolesController.java
@@ -1,7 +1,7 @@
 package ba.unsa.etf.nwtcinemamovies.controllers;
 
-import ba.unsa.etf.nwtcinemamovies.models.Movie;
-import ba.unsa.etf.nwtcinemamovies.services.MovieService;
+import ba.unsa.etf.nwtcinemamovies.models.UserRole;
+import ba.unsa.etf.nwtcinemamovies.services.UserRoleService;
 import ba.unsa.etf.nwtcinemamovies.utils.JSONConverter;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestBody;
@@ -12,36 +12,36 @@ import org.springframework.web.bind.annotation.RestController;
 import javax.transaction.Transactional;
 
 @RestController
-@RequestMapping(value = "movies", produces = "application/json")
-public class Movies extends AbstractController<MovieService> {
+@RequestMapping(value = "userRoles", produces = "application/json")
+public class UserRolesController extends AbstractController<UserRoleService> {
 
 	@Transactional
 	@RequestMapping(value = "create", method = RequestMethod.POST)
-	public void create(@RequestBody final Movie movie) {
-		service.save(movie);
+	public void create(@RequestBody final UserRole role) {
+		service.save(role);
 	}
 
 	@Transactional
 	@RequestMapping(value = "update", method = RequestMethod.POST)
-	public String update(@RequestBody final Movie movie) {
-		return JSONConverter.toJSON(service.update(movie));
+	public String update(@RequestBody final UserRole role) {
+		return JSONConverter.toJSON(service.update(role));
 	}
 
 	@Transactional
-	@RequestMapping(value = "{movieId}", method = RequestMethod.GET)
-	public String findById(@PathVariable("movieId") final Integer movieId) {
-		return JSONConverter.toJSON(service.findById(Movie.class, movieId));
+	@RequestMapping(value = "{roleId}", method = RequestMethod.GET)
+	public String findById(@PathVariable("roleId") final Integer roleId) {
+		return JSONConverter.toJSON(service.findById(UserRole.class, roleId));
 	}
 
 	@Transactional
 	@RequestMapping(value = "findAll", method = RequestMethod.GET)
 	public String findAll() {
-		return JSONConverter.toJSON(service.findAll(Movie.class));
+		return JSONConverter.toJSON(service.findAll(UserRole.class));
 	}
 
 	@Transactional
 	@RequestMapping(value = "delete", method = RequestMethod.DELETE)
-	public void delete(@RequestBody final Movie movie) {
-		service.delete(movie);
+	public void delete(@RequestBody final UserRole role) {
+		service.delete(role);
 	}
 }
